Refactored some file names
